+++ src/leiningen/new/lein_compojure.clj
@@ -24,4 +24,5 @@
              ["scripts/run-tmux" (render "run-tmux" data)]
              ["src/{{sanitized}}/env.clj" (render "env.clj" data)]
              ["src/{{sanitized}}/handler.clj" (render "handler.clj" data)]
+             ["src/{{sanitized}}/templating.clj" (render "templating.clj" data)]
              ["test/{{sanitized}}/handler_test.clj" (render "handler_test.clj" data)])))